requirements.txt
requests>=2.25.1
Pillow>=9.0.0
Flask>=3.0.0
flask-cors>=4.0.0
